telethon-ingest/requirements.txt
# Telethon Ingest service dependencies
# Telegram API
telethon==1.34.0
cryptg==0.4.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis and event streaming
redis==5.0.1
aioredis==2.0.1

# HTTP client
httpx==0.25.2
aiohttp==3.9.1

# Monitoring and metrics
prometheus-client==0.19.0
structlog==23.2.0

# Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
python-multipart==0.0.6
qrcode[pil]==7.4.2
Pillow==10.1.0

# Cryptography
cryptography==41.0.7

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Logging and debugging
rich==13.7.0
loguru==0.7.2